movable particles

+++ src/sample/IMoveController.java
@@ -3,7 +3,7 @@ package sample;
 public interface IMoveController {
     void moveForward ( int cm );
     void moveBackward ( int cm);
-    void turnLeft ( int angle );
-    void turnRight ( int angle );
+    void turnLeft ( double angle );
+    void turnRight ( double angle );
 
 }
